# Core support for enums, take #2