… presto-common/pom.xml
… presto-common/src/main/java/com/facebook/presto/common/type/TypeSignature.java
… presto-common/src/main/java/com/facebook/presto/common/type/VarcharType.java
… presto-common/src/test/java/com/facebook/presto/common/type/BenchmarkTypeSignatureParsing.java
… presto-common/src/test/java/com/facebook/presto/common/type/TestTypeSignature.java
… presto-docs/src/main/sphinx/develop/functions.rst
… presto-main/src/main/java/com/facebook/presto/metadata/SignatureBinder.java
… presto-main/src/main/java/com/facebook/presto/operator/annotations/FunctionsParserHelper.java
… presto-main/src/main/java/com/facebook/presto/sql/analyzer/ExpressionAnalyzer.java
… presto-main/src/main/java/com/facebook/presto/sql/analyzer/ExpressionTreeUtils.java
… presto-main/src/main/java/com/facebook/presto/sql/planner/ExpressionInterpreter.java
… presto-main/src/main/java/com/facebook/presto/sql/planner/RelationPlanner.java
… presto-main/src/main/java/com/facebook/presto/sql/planner/TranslationMap.java
… presto-main/src/main/java/com/facebook/presto/sql/relational/SqlToRowExpressionTranslator.java
… presto-main/src/main/java/com/facebook/presto/type/EnumCasts.java
… presto-main/src/main/java/com/facebook/presto/type/LongEnumOperators.java
… presto-main/src/main/java/com/facebook/presto/type/VarcharEnumOperators.java
… presto-main/src/main/java/com/facebook/presto/util/JsonUtil.java
… presto-main/src/test/java/com/facebook/presto/sql/analyzer/TestAnalyzer.java
… presto-main/src/test/java/com/facebook/presto/sql/query/TestJoinUsing.java
… presto-parser/src/main/java/com/facebook/presto/sql/tree/AstVisitor.java
… presto-parser/src/main/java/com/facebook/presto/sql/tree/EnumLiteral.java
… presto-spi/src/main/java/com/facebook/presto/spi/function/TypeParameter.java
… presto-spi/src/main/java/com/facebook/presto/spi/function/TypeVariableConstraint.java
… presto-tests/src/main/java/com/facebook/presto/tests/AbstractTestQueries.java
… presto-tests/src/main/java/com/facebook/presto/tests/TestingPrestoClient.java
… presto-tests/src/test/java/com/facebook/presto/tests/TestEnums.java