Second attempt to merge #14927 

This was previously reverted because of a performance regression detected in `TypeSignature.parseTypeSignature`

This regression is addressed in a newly added commit, which includes a benchmark.

Performance stats on BenchmarkTypeSignatureParsing are as follows on my machine:


```
parseRowTypeSignature
==================
* with enum parsing, unoptimized parsing: 600ms/op
* with enum parsing, optimized parsing: 2ms/op
* without enum parsing: 2ms/op

parseRowTypeSignatureWithEnums
==========================
* with enum parsing, optimized parsing: 2ms/op
```