@daniel-ohayon 2020-09-24
@rongrong I also added another new commit to the stack: Move enum parametric types to presto-common
This is so that plugins don't need to depend on presto-main to provide enum types.

I could also blend the last 3 commits into the existing commit but I figured that since the stack had already been merged into master and reverted, it might be clearer to leave the fixes as separate commits?