@rongrong 2020-09-24
> @rongrong I also added another new commit to the stack: Move enum parametric types to presto-common
> This is so that plugins don't need to depend on presto-main to provide enum types.
> 
> I could also blend these last 3 "fix" commits into the earlier commits but I figured that since the stack had already been merged into master and reverted, it might be clearer to leave the fixes as separate commits?

We already reverted the previous commits so let's just try to go clean and make each commit as bug-free as possible. It's easier to review the changes this way, but when we are ready to merge, please merge the new commits with the original ones introducing the change.