@daniel-ohayon 2020-09-24
OK, I combined the last 3 commits into the earlier ones (that includes your fix to the TreeRewriter)