@rongrong 2020-09-24
@yingsu00 @caithagoras We fixed the regression and added a benchmark to test `parseTypeSignature`. What do you guys fill about this? Should we do a verifier test before merging or rely on the nightly and release verification? I guess I'm the next release oncall so might be alright to just merge now? 😂